实现正确读取一个没有字段没有方法的类A

+++ src/main/java/vm/parser/U4.java
@@ -5,9 +5,9 @@ package vm.parser;
  * @version 1.0 2016/3/27
  */
 public class U4 {
-    public long value;
+    public int value;
 
-    public static U4 of(long value){
+    public static U4 of(int value){
         U4 u4 = new U4();
         u4.value = value;
         return u4;
